Make autoload thread-safe (in progress)
Added autoloadingMap to RubyModule to keep an autoloading state. It
consists of a RubyThread which invokes autoloading and an Object
assigned to the autoload constant while autoloading.

While autoloading, constantMap keeps UNDEF for the constant and
autoloadingMap keeps the assigned object instead. Constant lookup checks
if the value in constantMap is UNDEF, and iif it's UNDEF, it checks
autloadingMap as well.

It's still in development. For now, it passes
org.jruby.test.ScriptTestSuite but org.jruby.test.JRubyTestSuite causes
infinite loop. I'll check later.

Conflicts:

	src/org/jruby/RubyKernel.java
	src/org/jruby/RubyModule.java