Avoid infinite loop constant lookup
RubyModule#getConstantFromNoConstMissing() did autoload lookup when the
defined constant was UNDEF, then tried to detect the defined value at
the next loop. It caused infinite loop since the constantMap holds UNDEF
while autoloading in new thread-safe autoloading logic.

Now resolveUndefConstant() blocks if other thread is autoloading, and
returns correct value so making it simply return the given value from
resolveUndefConstant(). I checked all resolveUndefConstant() occurrences
in JRuby project and found there's no other loops.

ant test passes!

ant spec fails with 3 specs but it's from the intended behavir change
that Module#autoload does not remove the constant from constant table if
the loaded files does not define it now. The behavior is as same as
CRuby 1.9, and needed for autoload thread-safety. CRuby 1.8 removes the
constant, and it's hard to be thread-safe as far as it keeps the current
behavior.

I believe this JRuby/CRuby behavior incompatibility on 1.8 mode is OK to
introduce since the behavior is hard to imagine before trying it so no
one would depend on the behavior intentionally.

Conflicts:

	src/org/jruby/RubyModule.java